Split TLS observed config tests into OCP and HyperShift suites
Refactor TLS observed config tests to separate standalone OCP and
HyperShift test suites for independent ownership. Move shared types,
data, and test implementations to the parent tls package, with
platform-specific test registrations in dedicated subdirectories
(tls/ocp and tls/hypershift) each with their own OWNERS files.
Move generic helpers to test/extended/util for reuse.

Suite names updated:
- openshift/tls-observed-config-ocp (standalone OCP)
- openshift/tls-observed-config-hypershift (HyperShift)

Made-with: Cursor
