+++ test/extended/tls/hypershift/OWNERS
@@ -1,0 +1,13 @@
+# See the OWNERS docs: https://git.k8s.io/community/contributors/guide/owners.md
+
+reviewers:
+  - csrwng
+  - celebdor 
+  - jiezhao16 
+  - mehabhalodiya 
+
+approvers:
+  - csrwng
+  - celebdor 
+  - jiezhao16 
+  - mehabhalodiya
